add global and modular css
global css works just like plain css, you can use it in your JSX as: className= "class". Import the global.css in the _app.js file and it can be used anywhere in the project.
modular css is different. It works for a particular module only. Import it as: import styles from '../styles/x.module.css' and use it as: className={styles.class}